dit is allemaal fantastich

+++ include/CreateSolveLP.h
@@ -11,7 +11,7 @@
 
 
 //This function has input of the graph, the value y of the variables, the costs of the edges, the requirement on V*V and an edge to fix and a bool to say whether something needs to be fixed or not
-void solve_lp(lemon::ListGraph *,lemon::ListGraph::EdgeMap<double> *,lemon::ListGraph::EdgeMap<int> *, RequirementFunction*, lemon::ListGraph::EdgeMap<int> *, lemon::ListGraph::Edge *, double*, bool );    
+void solve_lp(lemon::ListGraph *,lemon::ListGraph::EdgeMap<double> *,lemon::ListGraph::EdgeMap<double> *, RequirementFunction*, lemon::ListGraph::EdgeMap<int> *, lemon::ListGraph::Edge *, double*, bool );    
 
 // it should solve the LP and adjust y to represent the changed value of the variables after solving the LP
 // Also it should occasionally fix some variables to try if the value is the same
